+++ sql/hive/src/main/scala/org/apache/spark/sql/hive/HiveSessionState.scala
@@ -87,8 +87,11 @@ class HiveSessionStateBuilder(session: SparkSession, parentState: Option[Session
 
     override val postHocResolutionRules: Seq[Rule[LogicalPlan]] =
       new DetermineTableStats(session) +:
-      catalog.ParquetConversions +:
-      catalog.OrcConversions +:
+      // These 2 rules must be run before all other DDL post-hoc resolution rules, i.e.
+      // `PreprocessTableCreation`, `PreprocessTableInsertion`, `DataSourceAnalysis` and
+      // `HiveAnalysis`.
+      ParquetConversions(session) +:
+      OrcConversions(session) +:
       PreprocessTableCreation(session) +:
       PreprocessTableInsertion(conf) +:
       DataSourceAnalysis(conf) +:
